Debug-Release/32-64-bits
Projetos configurados para versões Debug e Release nas plataformas de 32 e 64 bits

+++ Labs/Lab09/Galaga/Galaga/Renderer.cpp
@@ -443,7 +443,7 @@ void Renderer::Render()
         { return a->depth > b->depth; } );
 
     // quantidades de sprites a serem renderizados
-    uint spriteVectorSize = spriteVector.size();
+    uint spriteVectorSize = uint(spriteVector.size());
 
     if (spriteVectorSize == 0)
         return;
